requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pandas==1.3.5
numpy==1.21.6
openai==0.28.1
plotly==5.15.0
python-dotenv==0.21.1
gunicorn==20.1.0 